+++ includes/function.inc.sh
@@ -94,13 +94,21 @@ function welcome_short
 function welcome
 {
   welcome_short
-  echo "  * PHP 5.3-7.0 + Nginx + MariaDB installer"
+  echo "  * PHP 5.3-7.2 + Nginx + MariaDB installer"
   echo
 }
 
 function options
 {
   printf "  - Install ${GREEN}EPEL repo${NO_COLOR} / http://fedoraproject.org/wiki/EPEL\n"
+
+  if [ $PHP72 = "1" ]; then
+    printf "  - Install ${GREEN}PHP 7.2${NO_COLOR} from Remi repo / http://rpms.famillecollet.com/\n"
+  fi
+
+  if [ $PHP71 = "1" ]; then
+    printf "  - Install ${GREEN}PHP 7.1${NO_COLOR} from Remi repo / http://rpms.famillecollet.com/\n"
+  fi
 
   if [ $PHP70 = "1" ]; then
     printf "  - Install ${GREEN}PHP 7.0${NO_COLOR} from Remi repo / http://rpms.famillecollet.com/\n"